+++ Client/SizeFilter/SizeFilter.py
@@ -1,5 +1,7 @@
 import os
 
+import imagesize
+
 
 def check_size_in_KB(path):
     size_in_b = os.path.getsize(path)
@@ -8,7 +10,7 @@ def check_size_in_KB(path):
 
 
 def get_comparator(comparator, threshold=0):
-    return {"==": lambda reference, checked: abs(reference - checked) < threshold,
+    return {"==": lambda reference, checked: abs(reference - checked) <= threshold,
             ">": lambda reference, checked: reference > checked,
             ">=": lambda reference, checked: reference >= checked,
             "<": lambda reference, checked: reference < checked,
@@ -20,5 +22,15 @@ def filter_by_KB(paths, reference, comparator, threshold):
     comparator = get_comparator(comparator, threshold)
     for path in paths:
         if comparator(reference, check_size_in_KB(path)):
+            filtered_paths.append(path)
+    return filtered_paths
+
+
+def filter_by_pixels(paths, reference, comparator, threshold):
+    filtered_paths = []
+    comparator = get_comparator(comparator, threshold)
+    for path in paths:
+        width, height = imagesize.get(path)
+        if comparator(reference[0], width) and comparator(reference[1], height):
             filtered_paths.append(path)
     return filtered_paths
